+++ src/solidlsp/ls_process.py
@@ -2,7 +2,6 @@ import asyncio
 import json
 import logging
 import os
-import platform
 import socket
 import subprocess
 import threading
@@ -33,7 +32,7 @@ from solidlsp.lsp_protocol_handler.server import (
     make_request,
     make_response,
 )
-from solidlsp.util.subprocess_util import quote_arg, subprocess_kwargs, terminate_process_tree_with_kill_fallback
+from solidlsp.util import subprocess_util
 
 log = logging.getLogger(__name__)
 
@@ -459,14 +458,9 @@ class StdioLanguageServer(LanguageServerInterface):
         child_proc_env = os.environ.copy()
         child_proc_env.update(self.process_launch_info.env)
 
-        cmd = self.process_launch_info.cmd
-        is_windows = platform.system() == "Windows"
-        if not isinstance(cmd, str) and not is_windows:
-            # Since we are using the shell, we need to convert the command list to a single string
-            # on Linux/macOS
-            cmd = " ".join(map(quote_arg, cmd))
+        cmd = subprocess_util.convert_shell_cmd(self.process_launch_info.cmd)
         log.info("Starting language server process via command: %s", self.process_launch_info.cmd)
-        kwargs = subprocess_kwargs()
+        kwargs = subprocess_util.subprocess_kwargs()
         kwargs["start_new_session"] = self.start_independent_lsp_process
         self.process = subprocess.Popen(
             cmd,
@@ -513,7 +507,9 @@ class StdioLanguageServer(LanguageServerInterface):
                 log.debug(f"Exception during graceful shutdown: {e}")
                 # Ignore errors here, we are proceeding to terminate anyway.
             # terminate the process
-            terminate_process_tree_with_kill_fallback(self.process, terminate_timeout=timeout, process_name=f"LS[{self.language.value}]")
+            subprocess_util.terminate_process_tree_with_kill_fallback(
+                self.process, terminate_timeout=timeout, process_name=f"LS[{self.language.value}]"
+            )
         finally:
             self.process = None